fix: refactor the postgresql and duckdb config management
